Move transform interpreter utils to dialect.
This is a preparatory commit for fixing the handling of multiple
transform library modules. The plan is to merge every newly loaded
module into the global library module rather than keeping around a list.
Since that merging happens in the dialect and depends on the dialect,
the two depend on each other and thus have to live in the same CMake
target.
